% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonjour.R
\name{bonjour}
\alias{bonjour}
\title{La fonction qui dit bonjour}
\usage{
bonjour(couleur = "pink", fond = "limegreen")
}
\arguments{
\item{couleur}{la couleur du texte}

\item{fond}{la couleur du fond}
}
\value{
Un graphe simple
}
\description{
"Bonjour !" : Cette fonction est un exemple de fonction graphique avec deux paramètres simples : la couleur du texte et la couleur du fond.
}
\examples{
bonjour()
}
