% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonjour.R
\docType{data}
\name{features}
\alias{features}
\title{Exemple de données de méthylation (features)}
\format{
GRanges
}
\description{
Ces données de méthylation ont été sauvées sous une forme plus pratique pour le cours. features est un objet de classe GRanges qui contient l'information de deux gènes,
}
\references{
\url{https://bioconductor.org/packages/release/bioc/vignettes/trackViewer/inst/doc/dandelionPlot.html}
}
\keyword{data}
