% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonjour.R
\docType{data}
\name{gr}
\alias{gr}
\title{Exemple de données de méthylation (gr)}
\format{
GRanges
}
\description{
Ces données de méthylation ont été sauvées sous une forme plus pratique pour le cours. gr est un objet de classe GRanges qui contient la position sur le gène TYMP à laquelle le graphe doit commencer.
}
\references{
\url{https://bioconductor.org/packages/release/bioc/vignettes/trackViewer/inst/doc/dandelionPlot.html}
}
\keyword{data}
