% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonjour.R
\docType{data}
\name{methy}
\alias{methy}
\title{Exemple de données de méthylation (méthylation)}
\format{
GRanges
}
\usage{
methy
}
\description{
Ces données de méthylation ont été sauvées sous une forme plus pratique pour le cours. methy est un objet de classe GRanges qui contient des sites de méthylation,
}
\references{
\url{https://bioconductor.org/packages/release/bioc/vignettes/trackViewer/inst/doc/dandelionPlot.html}
}
\keyword{data}
