% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyViz.R
\docType{package}
\name{tidyViz-package}
\alias{tidyViz}
\alias{tidyViz-package}
\title{Package pour la formation EBAII
Contient les vignettes (les slides de cours), un jeu de données et les tutoriels}
\description{
Ce package contient les slides du cours (sous forme de vignettes)
  ainsi que des tutoriels développés avec learnr.
}
\keyword{internal}
