% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonjour.R
\docType{data}
\name{fruits}
\alias{fruits}
\title{Le jeu de données fruits}
\description{
Composition nutritionnelle de 51 composés alimentaires issus du groupe "fruits" défini dans les données Ciqual de 2020.
}
\details{
La table fruits est de classe "tibble" et contient les colonnes suivantes : 
\itemize{
 \item nom : Nom du fruit
 \item groupe : Groupe de fruit
 \item Energie : Energie, (kJ/100 g)
 \item Eau : Eau (g/100 g)
 \item Proteines : Protéines, N x 6.25 (g/100 g)
 \item Glucides : Glucides (g/100 g)
 \item Lipides : Lipides (g/100 g)
 \item Sucres : Sucres (g/100 g)
 \item Fructose : Fructose (g/100 g)
 \item Fibres : Fibres alimentaires (g/100 g)
 \item Calcium : Calcium (mg/100 g)
 \item Magnesium : Magnésium (mg/100 g)
 \item Phosphore : Phosphore (mg/100 g)
 \item Potassium : Potassium (mg/100 g)
 \item Zinc : Zinc (mg/100 g)
 \item BetaCarotene : Beta-Carotène (µg/100 g)
 \item VitamineE : Vitamine E (mg/100 g)
 \item VitamineC : Vitamine C (mg/100 g)
}
}
\references{
\url{https://ciqual.anses.fr/}
}
\keyword{data}
